Youtube_chatbot/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
huggingface-hub
youtube-transcript-api
python-dotenv
faiss-cpu
